Moved config to common
